requirements.txt
requests==2.32.3
python-dotenv==1.0.1
python-dateutil==2.9.0.post0